pyproject.toml
[tool.poetry]
name = "mba-ia-desafio-ingestao-busca"
version = "0.1.0"
description = ""
authors = ["MieleSantos <mielesnts@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
langchain-openai = "^1.1.7"
langchain-google-genai = "^4.1.3"
langchain-postgres = "^0.0.16"
langchain-community = "^0.4.1"
python-dotenv = "^1.2.1"
pypdf = "^6.6.0"
pymupdf = "^1.26.7"
loguru = "^0.7.2"
psycopg2-binary = "^2.9.10"
langchain = "^1.2.6"


[tool.poetry.group.dev.dependencies]
mypy = "^1.19.1"
black = "^26.1.0"
isort = "^7.0.0"
invoke = "^2.2.1"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
